Comments from Jon
The whole thing was really simple to follow, and you did a nice job in
avoiding complicating the design. This helped me find issues much more
quickly than in other codebases!

Other than a few checks that need to be added, the biggest piece to
change is the price calculation. As it stands, it's extremely easy to
overflow the calculation, which might make it possible for people to
exploit.

I've provided a proptest that you can use to make sure that the function
is safe. You can run it with `cargo test` from `programs/gg`. As a quick
solution, you can cast everything to `u128`s in the function, and then
convert back to a `u64` at the end. If it fails, you return an error.

Also, you can use checked math. If the checked math ever returns `None`,
you return an error. This will ensure that the program can never get
hacked, even with huge input numbers.